연락처정보관련함수 분리
