+++ app/src/main/java/com/capstone/belink/UIActivity/FriendSettingActivity.kt
@@ -11,6 +11,8 @@ import android.view.View
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import com.capstone.belink.Model.User
+import com.capstone.belink.Utils.getStringArrayPref
+import com.capstone.belink.Utils.setStringArrayPref
 import com.capstone.belink.databinding.ActivityFriendSettingBinding
 import org.json.JSONArray
 import org.json.JSONException
@@ -33,7 +35,7 @@ class FriendSettingActivity : AppCompatActivity() {
         }
 
         binding.btnGetContact.setOnClickListener {
-            getStringArrayPref("contact")
+            getStringArrayPref(this,"contact")
         }
     }
 
@@ -58,7 +60,9 @@ class FriendSettingActivity : AppCompatActivity() {
             contactList.add(obj)
         }
         contacts.close()
-        setStringArrayPref("contact",contactList)
+
+        setStringArrayPref(this,"contact",contactList)
+
         Toast.makeText(this, "연락처 정보를 가져왔습니다.", Toast.LENGTH_SHORT).show()
     }
 
@@ -66,52 +70,6 @@ class FriendSettingActivity : AppCompatActivity() {
     override fun onDestroy() {
         mBinding=null
         super.onDestroy()
-    }
-
-    fun setStringArrayPref(key:String,values:MutableList<User>){
-        val pref:SharedPreferences=getSharedPreferences(key, MODE_PRIVATE)
-        val edit:SharedPreferences.Editor = pref.edit()
-        edit.putString(key,null)
-        val dataList=JSONArray()
-        for(i in values.indices){
-            val tempJsonObject = JSONObject()
-            Log.d("이름","${values[i].username}")
-            Log.d("전화번호","${values[i].phNum}")
-            tempJsonObject.put("username",values[i].username)
-            tempJsonObject.put("phNum",values[i].phNum)
-            dataList.put(tempJsonObject)
-        }
-        if(values.isNotEmpty()) {
-            edit.putString(key, dataList.toString())
-        }else{
-            edit.putString(key,null)
-        }
-        edit.apply()
-    }
-
-    fun getStringArrayPref(key:String):MutableList<User>{
-        val pref:SharedPreferences=getSharedPreferences(key, MODE_PRIVATE)
-        val edit:SharedPreferences.Editor = pref.edit()
-        val json=pref.getString(key,null)
-        var uri : MutableList<User> = ArrayList()
-        if(json != null){
-            try{
-                val temp = JSONArray(json)
-                for(i in 0 until temp.length()){
-                    val iObject = temp.getJSONObject(i)
-                    val username = iObject.getString("username")
-                    val phNum = iObject.getString("phNum")
-                    val obj = User(username=username,phNum = phNum)
-                    uri.add(obj)
-                    Log.d("$phNum","$username")
-                }
-            }catch (e:JSONException){
-                e.printStackTrace()
-            }
-        }
-        return uri
-
-
     }
